+++ api/graphql/resolver/cpanel/users.js
@@ -1,7 +1,7 @@
 import bcrypt from "bcrypt";
 import Users from "../../../../db/models/Users.js";
 import { generateToken } from "../../../../util/helper.js";
-
+import { validateUserInput } from "../../../../util/validateUserInput.js";
 import { UserInputError } from "apollo-server";
 
 const resolvers = {
@@ -12,7 +12,7 @@ const resolvers = {
           order: [['id', 'DESC']],
         });
         return users;
-      } catch (error) { 
+      } catch (error) {
         console.error("Error fetching users:", error);
         throw new Error("Failed to fetch users.");
       }
@@ -21,90 +21,33 @@ const resolvers = {
   Mutation: {
     createUser: async (_, { userInput }) => {
       try {
-        const {
-          firstname, lastname, email, password, role, company_name,
-          country_code, number, alt_country_code, alt_number, country, state, city, zipcode,
-          address1, address2, ult_parent_id, status, dob, no_of_calls, callback, gender, parent_id,
-        } = userInput;
-    
-        const errors = {};
-    
-        // Field validations
-        if (!firstname || firstname.trim() === "") {
-          errors.firstname = "First name is required.";
-        }
-        if (!email || email.trim() === "") {
-          errors.email = "Email is required.";
-        }
-    
-        if (!password || password.length < 6) {
-          errors.password = "Password must be at least 6 characters.";
-        }
-    
-        if (!company_name || company_name.trim() === "") {
-          errors.company_name = "Company name is required.";
-        }
-    
-        if (!number || number.toString().length < 7) {
-          errors.number = "Valid phone number is required.";
-        }
-    
-        if (!country || country.trim() === "") {
-          errors.country = "Country is required.";
-        }
-        if (!state || state.trim() === "") {
-          errors.state = "State is required.";
-        }
-        if (!city || city.trim() === "") {
-          errors.city = "City is required.";
-        }
-    
-        if (!address1 || address1.trim() === "") {
-          errors.address1 = "Address 1 is required.";
-        }
-    
+        // Validate user input using the helper function
+        const errors = validateUserInput(userInput);
+
         // Check if email already exists
+        const { email } = userInput;
         const existingUser = await Users.findOne({ where: { email } });
         if (existingUser) {
           errors.email = "Email is already in use.";
         }
-    
-        // If any errors exist, throw them
+
+        // If there are any validation errors, throw them
         if (Object.keys(errors).length > 0) {
           const firstField = Object.keys(errors)[0];
           throw new UserInputError(errors[firstField], { field: firstField });
-        }        
-    
+        }
+
         // Hash password
-        const hashedPassword = await bcrypt.hash(password, 10);
-    
+        const hashedPassword = await bcrypt.hash(userInput.password, 10);
+
         // Create new user
         const newUser = await Users.create({
-          firstname,
-          lastname,
-          email,
+          ...userInput,
           password: hashedPassword,
-          role,
-          company_name,
-          country_code,
-          number,
-          alt_country_code,
-          alt_number,
-          country,
-          state,
-          city,
-          zipcode,
-          address1,
-          address2,
-          ult_parent_id: ult_parent_id === "0" ? null : ult_parent_id,
-          parent_id: parent_id === "0" ? null : parent_id,
-          status,
-          dob,
-          callback,
-          no_of_calls,
-          gender
+          ult_parent_id: userInput.ult_parent_id === "0" ? null : userInput.ult_parent_id,
+          parent_id: userInput.parent_id === "0" ? null : userInput.parent_id,
         });
-    
+
         return {
           ...newUser.dataValues,
         };
@@ -116,6 +59,7 @@ const resolvers = {
         throw new Error(error.message || "Failed to create user.");
       }
     },
+
     login: async (_, { email, password }) => {
       try {
         // Check if user exists
@@ -147,9 +91,7 @@ const resolvers = {
         throw new Error(error.message || "Login failed.");
       }
     },
-    
-  }
-
+  },
 };
 
 export default resolvers;
